+++ src/Service/Tags/OrderTagService.php
@@ -27,9 +27,13 @@ class OrderTagService
      */
     private $tagRepository;
 
+    /**
+     * @param EntityRepository $orderRepository
+     * @param EntityRepository $tagRepository
+     */
     public function __construct(
-        EntityRepository $orderRepository,
-        EntityRepository $tagRepository
+         $orderRepository,
+         $tagRepository
     ) {
         $this->orderRepository = $orderRepository;
         $this->tagRepository = $tagRepository;