+++ src/Service/TransactionService.php
@@ -20,8 +20,10 @@ class TransactionService
 
     /**
      * Creates a new instance of the transaction service.
+     *
+     * @param mixed $orderTransactionRepository
      */
-    public function __construct(EntityRepository $orderTransactionRepository)
+    public function __construct($orderTransactionRepository)
     {
         $this->orderTransactionRepository = $orderTransactionRepository;
     }